improved render loop termination to catch circular dependencies
Instead of checking that the serializded data text before and after a render is the same, we now
generate and store a sha1 hash of the text. If any hash is repeated, we
terminate the loop. This garantees that we don't get caught in an
infinite loop caused by circular dependencies.