added support for user-defined filters
Some parsers require mako expressions to be inside of a string in order
to parse. However, if we need these parameters to be parsed as a number
so they can be used in other expressions.

In order to get around this, we where converting all values to numbers
(well, all values would convert to a number), but this was not great
because it may be that the user actually wants a number in a string.

Now the user can specify their own filter function, which by default is
a function that tires to convert everything to a number.